ArmStub Implementation
