+++ src/main/deploy/robotconfig/00802F17DEE0.json
@@ -4,7 +4,8 @@
 
     "capabilities" : {
         "hasInvertedRoll": true,
-        "hasInvertedDriveEncoderSensorPhase": true
+        "hasInvertedDriveEncoderSensorPhase": true,
+        "hasArm": true
     },
 
     "drivetrain" : {
