+++ src/main/java/bhs/devilbotz/RobotContainer.java
@@ -24,6 +24,7 @@ import bhs.devilbotz.lib.AutonomousModes;
 import bhs.devilbotz.subsystems.Arm;
 import bhs.devilbotz.subsystems.DriveTrain;
 import bhs.devilbotz.subsystems.Gripper;
+import bhs.devilbotz.subsystems.stubs.ArmStub;
 import bhs.devilbotz.utils.ShuffleboardManager;
 import com.pathplanner.lib.PathConstraints;
 import com.pathplanner.lib.PathPlanner;
@@ -56,7 +57,7 @@ public class RobotContainer {
 
   private final Gripper gripper = new Gripper();
 
-  private final Arm arm = new Arm();
+  private Arm arm;
 
   private final ShuffleboardManager shuffleboardManager = new ShuffleboardManager();
 
@@ -75,6 +76,12 @@ public class RobotContainer {
 
   /** The container for the robot. Contains subsystems, OI devices, and commands. */
   public RobotContainer() {
+    if (Robot.checkCapability("hasArm")) {
+      arm = new Arm();
+    } else {
+      arm = new ArmStub();
+    }
+
     // Configure the trigger bindings
     configureBindings();
